+++ src/TomsToolbox.Avalonia/GlobalUsings.cs
@@ -1,0 +1,3 @@
+﻿global using Dispatcher = Avalonia.Threading.Dispatcher;
+global using DispatcherPriority = Avalonia.Threading.DispatcherPriority;
+global using DispatcherTimer = Avalonia.Threading.DispatcherTimer;